Finished Genetic Algorithm :-)
Run it from Referee_test.py
enter 20 at the first prompt (how many genes in a genome?)
enter the number of genomes in the population
then use the playGame() function to start the game.

Each Mean listed represents the mean score for each generation as it
evolves.  the algorithm will end  when it detects the optimized winning
strategy.