NKmodel
the finished version. takes any gene length and k value, creates a
fitness landscape. landscape is the mean of the fitness of each alleal
based on the values of the others (not the sum), but this is easily
changed. run it N times and it records successes. last best gene is
modified to get the next.